Load png into 16bit depth as to not lose precision from sRGB -> linear

+++ twil/loader/png.hpp
@@ -11,7 +11,7 @@ namespace loader {
 class Png
 {
 private:
-	std::unique_ptr<std::uint8_t[]> bytes_;
+	std::unique_ptr<std::uint16_t[]> bytes_;
 	std::uint16_t width_;
 	std::uint16_t height_;
 
@@ -20,8 +20,8 @@ public:
 
 	std::uint16_t GetWidth() const;
 	std::uint16_t GetHeight() const;
-	std::uint8_t const * begin() const;
-	std::uint8_t const * end() const;
+	std::uint16_t const * begin() const;
+	std::uint16_t const * end() const;
 };
 
 }
